refactor: extract response_matches_json_schema
move helper functions to helpers.py
