+++ examples/team_recommender/tests/example_9_threshold/test_measurement_is_within_threshold.py
@@ -1,17 +1,14 @@
 import json
-from pathlib import Path
 from typing import List
 
 import openai
-import pytest
-from helpers import load_json_fixture, natural_sort_key
-from jsonschema import FormatChecker, validate
+from helpers import is_within_expected, load_json_fixture
 from openai import OpenAI
 from openai.types.chat.chat_completion import Choice
+from response_matches_json_schema import response_matches_json_schema
 from retry import retry
 from settings import root_dir
 
-from cat_ai import analyse_sample_from_test
 from cat_ai.reporter import Reporter
 from cat_ai.runner import Runner
 
@@ -28,95 +25,12 @@ def get_developer_names_from_response(response) -> set[str]:
     return {developer["name"] for developer in response["developers"]}
 
 
-def response_matches_json_schema(response: str, schema: any) -> bool:
-    """
-    Validates if a given response matches the provided JSON schema.
-
-    :param response: The response JSON data as a string.
-    :param schema: The schema to validate against.
-    :return: True if the response matches the schema, otherwise False.
-    """
-    try:
-        validate(instance=response, schema=schema, format_checker=FormatChecker())
-        return True
-    except Exception as e:
-        print(f"An unexpected error occurred: {e}")
-        return False
-
-
 def test_response_matches_json_schema():
     # Load example output and schema
     example_output = load_json_fixture("example_output.json")
     schema = load_json_fixture("output_schema.json")
 
     assert response_matches_json_schema(example_output, schema)
-
-
-def is_within_a_range(value, left, right):
-    return left <= value <= right
-
-
-def test_is_within_expected():
-    assert is_within_expected(0.8, 0, 5)
-    assert is_within_expected(0.8, 2, 5)
-    assert not is_within_expected(0.8, 3, 5)
-    assert not is_within_expected(0.8, 4, 5)
-    assert not is_within_expected(0.8, 5, 5)
-    assert is_within_expected(0.8, 26, 100)
-    assert is_within_expected(0.8, 14, 100)
-    assert is_within_expected(0.97, 1, 2)
-    small_size_warning = "after measuring 2x 100 runs and getting 3 failures"
-    assert is_within_expected(0.97, 0, 1), small_size_warning
-
-
-def is_within_expected(success_rate: float, failure_count: int, sample_size: int) -> bool:
-    if sample_size <= 1:
-        return True
-    success_portion = int(success_rate * sample_size)
-    success_analysis = analyse_sample_from_test(success_portion, sample_size)
-    return is_within_a_range(
-        sample_size - failure_count,
-        success_analysis.confidence_interval_count[0],
-        success_analysis.confidence_interval_count[1],
-    )
-
-
-def test_success_rate():
-    tiny_set_analysis = analyse_sample_from_test(1, 2)
-    assert tiny_set_analysis.proportion == 0.5
-    interval = tiny_set_analysis.confidence_interval_prop
-
-    assert interval[0] <= 0 and interval[1] >= 1, (
-        f"interval includes all possible values: {interval} does not contain [0, 1]"
-    )
-
-
-def test_sort_names_with_numbers():
-    unsorted = [
-        "example_1_text_response",
-        "example_10_threshold",
-        "example_2_unit",
-        "example_8_retry_network",
-        "example_9_retry_with_open_telemetry",
-    ]
-    assert [
-        "example_10_threshold",
-        "example_1_text_response",
-        "example_2_unit",
-        "example_8_retry_network",
-        "example_9_retry_with_open_telemetry",
-    ] == sorted(unsorted), "The list should be sorted by the number in the name"
-
-    correctly_sorted = [
-        "example_1_text_response",
-        "example_2_unit",
-        "example_8_retry_network",
-        "example_9_retry_with_open_telemetry",
-        "example_10_threshold",
-    ]
-    assert sorted([Path(p) for p in unsorted], key=natural_sort_key) == [
-        Path(p) for p in correctly_sorted
-    ], "example_10_threshold should be last, while example_1_text_response should be first"
 
 
 def test_metrics_within_range():
@@ -171,67 +85,6 @@ def test_metrics_within_range():
     assert is_within_expected(expected_success_rate_measured, failure_count, sample_size), (
         f"Expected {expected_success_rate_measured} to be within of the success rate"
     )
-
-
-def process_row(row: tuple[int, int, float]) -> str:
-    return f"{row[0]} failures out of {row[1]} is within {row[2] * 100:.0f}% success rate"
-
-
-@pytest.mark.parametrize(
-    "row",
-    [(1, 5, 0.97), (2, 5, 0.95), (6, 100, 0.97), (15, 100, 0.80), (27, 100, 0.80)],
-    ids=lambda row: process_row(row),
-)
-def test_success_rate_is_within_expected_error_margin_with_90_percent_confidence(
-    assert_success_rate, row
-):
-    failure_count, total_test_runs, expected_rate = row
-    results = generate_examples(failure_count, total_test_runs)
-    assert_success_rate(results, expected_rate)
-
-
-def generate_examples(failure_count, total_test_runs):
-    return failure_count * [False] + (total_test_runs - failure_count) * [True]
-
-
-@pytest.mark.parametrize(
-    "row",
-    [
-        (
-            1,
-            10,
-            0.70,
-            [
-                "New Success rate 0.900 with 90% confidence exceeds expected: 0.7",
-                "Expecting: 0.744 <= 0.700 <= 1.056",
-                "Got: expected=0.7 <= analysis.lower_interval=0.74",
-            ],
-        ),
-        (
-            1,
-            1000,
-            0.98,
-            [
-                "New Success rate 0.999 with 90% confidence exceeds expected: 0.98",
-                "Expecting: 0.997 <= 0.980 <= 1.001",
-                "Got: expected=0.98 <= analysis.lower_interval=0.997",
-            ],
-        ),
-    ],
-    ids=lambda row: row[-1][0],
-)
-def test_beyond_expected_success_rate(assert_success_rate, row):
-    failure_count, total_test_runs, expected_rate, success_messages = row
-    results = generate_examples(failure_count, total_test_runs)
-    with pytest.raises(AssertionError) as excinfo:
-        assert_success_rate(results, expected_rate)
-
-    message = str(excinfo.value)
-    for expected_message in success_messages:
-        assert expected_message in message
-    assert "Expecting: " in message
-    assert "Got: expected=0" in message
-    assert "<= analysis.lower_interval=0." in message
 
 
 @retry(
